+++ src/pages/zone/index.tsx
@@ -36,6 +36,7 @@ import { useHistory } from 'react-router-dom';
 import { FaSuitcase } from 'react-icons/fa';
 import MobileSortDialog from "../../components/MobileSortDialog"
 import MobileFilterDialog from "../../components/MobileFilterDialog"
+import { camelCase } from 'lodash';
 
 
 
@@ -131,6 +132,7 @@ const intialState = {
 };
 
 const Zone = () => {
+  const renderedFrom = camelCase(routes?.zone.title)
   const history = useHistory();
   const location = useLocation();
   const toastConfig = useContext(CustomToastContext);
@@ -158,7 +160,7 @@ const Zone = () => {
   const columnState = JSON.parse(localStorage.getItem('productCategoryPage'));
   const [isAllChecked, setIsAllChecked] = useState(false);
   const [clonedData, setClonedData] = useState([]);
-  const localStorageSelectedRecords = `${routes.zone.title}_selected`;
+  const localStorageSelectedRecords = `${renderedFrom}_selected`;
   const [isOpenDialog, setisOpenDialog] = useState(false)
   const [sortOpen, setSortOpen] = useState(false);
   const [anchorEl, setAnchorEl] = useState(null);
@@ -214,7 +216,7 @@ const Zone = () => {
                 { field: o?.fieldData?.fieldName, headerName: o?.fieldData?.fieldLabel, show: true, disabled: true, cellRenderer: 'nameRenderer' }
               ];
             } else {
-              let currentColumn = getColumnData(routes.zone.title, o?.fieldData, routes.zone.path);
+              let currentColumn = getColumnData(renderedFrom, o?.fieldData, routes.zone.path);
 
               if (currentColumn !== null) {
                 columns = [...columns, currentColumn?.columnData];
@@ -639,7 +641,7 @@ const Zone = () => {
                 setZoneId(data.id);
                 setOpen({ open: true, isClone: true });
               }}
-              renderedFrom={routes.productCategory.title}
+              renderedFrom={renderedFrom}
             />
           ) : (
             <CustomAgGrid
@@ -654,7 +656,7 @@ const Zone = () => {
               page={page}
               allowAction={true}
               loading={loading}
-              renderedFrom={routes.productCategory.title}
+              renderedFrom={renderedFrom}
               refreshGrid={fetchZone}
             />
           )